adapt spin/wakeup for multiple MNs

+++ Sherman/src/mb_utils.cpp
@@ -410,7 +410,7 @@ int check_CN_correctness(
 	// }
 	uint64_t baseAddr = dsm->getBaseAddr();
 	uint64_t *long_data = (uint64_t *) baseAddr;
-	for (int i = 0; i < GB(dsmSize) / sizeof(uint64_t); i++) {
+	for (size_t i = 0; i < GB(dsmSize) / sizeof(uint64_t); i++) {
 		datasum += long_data[i];
 		// if (long_data[i] > 0) {
 		// 	cerr << i << ": " << long_data[i] << endl;
